Revert "Revert "ver-2.0""
This reverts commit 536260bb968ce9d514f10ecf8af585da640bd851.

+++ CodevsForStudent/AI.hpp
@@ -203,7 +203,7 @@ private:
 
 		int average = 0;
 		for (const auto& v : blockTop) average += v;
-		average /= blockTop.size();
+		average /= (int)blockTop.size();
 
 		//R’J‚ÌŒŸõEŒ`‚Ì•]‰¿
 		int flatScore = 0;
